+++ src/runtime/CMakeLists.txt
@@ -1,4 +1,4 @@
 target_sources(bayesmix
   PUBLIC
-    Factory.hpp
+    factory.hpp
 )
